Refactor content scripts: fix security, bugs, split modules, deduplicate
Security fixes:
- Replace shell-injectable execSync+curl with fetch() in page-improver SCRY search
- Add path sandboxing to page-improver read_file tool handler
- Move hardcoded SCRY API key to env var with fallback in both scripts

Bug fixes:
- Swap gap-fill/review phase ordering so gap-fill can use review results
- Fix validation phase to validate new content (write to file, restore after)
- Fix runSynthesis called with wrong quality args ('opus'→'quality')
- Update stale component refs (KeyPeople/KeyQuestions→R/DataInfoBox/DataExternalLinks)
- Remove dead validate-quick phase and dead sidebar coverage checks

Module split (page-creator.mjs: 2400→500 lines):
- Extract 9 modules under tooling/content/creator/
- duplicate-detection, canonical-links, research, source-fetching,
  synthesis, verification, validation, grading, deployment
- page-creator.mjs becomes thin CLI entry point

Deduplication:
- grade-content: use shared parseFrontmatter, findMdxFiles, callClaude
- grade-by-template: use shared countWords/countTables/countDiagrams
- Export counting functions from metrics-extractor.mjs
- validation.mjs: delegate to shared componentImportsRule
- Fix WIKI_COMPONENTS to match actual component exports
- Add missing mkdirSync for grade-content output directory

Tests:
- 21 tests for creator modules (duplicate detection, URL extraction)
- 23 tests for metrics-extractor exports
- All existing tests continue to pass
