+++ tooling/content/creator/index.mjs
@@ -1,0 +1,15 @@
+/**
+ * Page Creator Module Index
+ *
+ * Re-exports all sub-modules for use by the page-creator CLI entry point.
+ */
+
+export { checkForExistingPage } from './duplicate-detection.mjs';
+export { findCanonicalLinks } from './canonical-links.mjs';
+export { runPerplexityResearch, runScryResearch } from './research.mjs';
+export { registerResearchSources, fetchRegisteredSources, processDirections } from './source-fetching.mjs';
+export { runSynthesis } from './synthesis.mjs';
+export { runSourceVerification } from './verification.mjs';
+export { ensureComponentImports, runValidationLoop, runFullValidation } from './validation.mjs';
+export { runGrading } from './grading.mjs';
+export { createCategoryDirectory, deployToDestination, validateCrossLinks, runReview } from './deployment.mjs';
